CVS-114407 [ONNX FE] SplitToSequence and SequenceAt operations - Made mixed-sign comparisons safer in SequenceAt

+++ src/frontends/onnx/frontend/src/op/sequence_at.cpp
@@ -42,12 +42,11 @@ ov::OutputVector sequence_at(const ov::frontend::onnx::Node& node) {
 
     const auto position_value = position_const->cast_vector<std::int64_t>()[0];
 
-    const auto input_sequence_length = input_sequence->get_sequence().size();
+    const auto input_sequence_length = static_cast<std::int64_t>(input_sequence->get_sequence().size());
 
     const auto position_value_normalized = position_value < 0 ? position_value + input_sequence_length : position_value;
-    OPENVINO_ASSERT(
-        position_value_normalized >= 0 && position_value_normalized < static_cast<std::int64_t>(input_sequence_length),
-        "SequenceAt: 'position' is out of bounds");
+    OPENVINO_ASSERT(position_value_normalized >= 0 && position_value_normalized < input_sequence_length,
+                    "SequenceAt: 'position' is out of bounds");
 
     return {input_sequence->get_sequence().at(position_value_normalized)};
 }
